+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @seed-design/mcp
 
+## 0.1.15
+
+### Patch Changes
+
+- @seed-design/figma@0.1.15
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [f806356]
+- Updated dependencies [1982494]
+  - @seed-design/figma@0.1.14
+
+## 0.1.13
+
+### Patch Changes
+
+- @seed-design/figma@0.1.13
+
+## 0.1.12
+
+### Patch Changes
+
+- @seed-design/figma@0.1.12
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [9993e0e]
+  - @seed-design/figma@0.1.11
+
 ## 0.1.10
 
 ### Patch Changes