+++ packages/rootage/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @seed-design/rootage-artifacts
 
+## 0.0.6
+
+### Patch Changes
+
+- 8299ba9: Snackbar 컴포넌트를 업데이트합니다.
+
+  - root 영역에 maxWidth 스펙을 추가합니다.
+  - `pauseOnInteraction`의 기본값을 `false`에서 `true`로 변경합니다.
+
+## 0.0.5
+
+### Patch Changes
+
+- f806356: Page Banner 컴포넌트를 추가합니다. Inline Banner 컴포넌트를 deprecate합니다.
+
+  - Inline Banner 컴포넌트 대비 모든 `tone`에서 모든 `variant`를 지원하며, 내부 Button의 충분한 터치 영역을 보장합니다.
+
+  ```tsx
+  <PageBanner
+    tone="informative"
+    variant="weak"
+    description="사업자 정보를 등록해주세요."
+    suffix={
+      <PageBannerButton asChild>
+        <a href="https://www.daangn.com" target="_blank" rel="noreferrer">
+          새 탭에서 열기
+        </a>
+      </PageBannerButton>
+    }
+  />
+  ```
+
+  시맨틱 색상 토큰을 추가하고 수정합니다.
+
+  - `$color.bg.positive-solid-pressed`: theme-dark에서 `$color.palette.green-500` → `$color.palette.green-600`
+  - `$color.bg.warning-solid-pressed` 추가
+
+- 1982494: Badge 컴포넌트를 업데이트합니다.
+
+  - `tone=warning` variant를 추가합니다.
+  - `maxWidth` 스펙을 추가합니다.
+
+  신규 시맨틱 색상 토큰을 추가합니다.
+
+  - `$color.fg.warning`
+  - `$color.stroke.warning`
+  - `$color.fg.brand-contrast`
+  - `$color.bg.brand-weak`
+  - `$color.bg.brand-weak-pressed`
+
+## 0.0.4
+
+### Patch Changes
+
+- 0be9b00: Avatar, Avatar Stack 컴포넌트에 `size=108` variant를 추가합니다.
+
 ## 0.0.3
 
 ### Patch Changes